Move template helper into helpers.py
