+++ ckanext/lacounts/plugin.py
@@ -2,31 +2,13 @@ import ckan.plugins as plugins
 import ckan.plugins.toolkit as toolkit
 from ckan.lib.plugins import DefaultTranslation
 
+from ckanext.lacounts.helpers import get_image_for_group
+
 
 class LacountsPlugin(plugins.SingletonPlugin, DefaultTranslation):
     plugins.implements(plugins.IConfigurer)
     plugins.implements(plugins.ITranslation)
     plugins.implements(plugins.ITemplateHelpers)
-
-    @staticmethod
-    def get_image_for_group(group_name):
-        '''
-        Render an inline svg snippet for the named group (topic). These groups
-        correlate with those created by the `create_featured_topics` command.
-        '''
-        jinja_env = toolkit.config['pylons.app_globals'].jinja_env
-        groups = {
-            'education': 'icons/book.svg',
-            'environment': 'icons/leaf.svg',
-            'health': 'icons/heart.svg',
-            'housing': 'icons/keys.svg',
-            'immigration': 'icons/passport.svg',
-            'transportation': 'icons/bus.svg'
-        }
-        svg = jinja_env.get_template(groups[group_name])
-        img = toolkit.literal("<span class='image %s'>" % group_name
-                              + svg.render() + "</span>")
-        return img
 
     # IConfigurer
     def update_config(self, config_):
@@ -36,4 +18,4 @@ class LacountsPlugin(plugins.SingletonPlugin, DefaultTranslation):
 
     # ITemplateHelpers
     def get_helpers(self):
-        return {'get_image_for_group': self.get_image_for_group}
+        return {'get_image_for_group': get_image_for_group}
